+++ modules/ui_definitions.py
@@ -558,7 +558,6 @@ def create_settings(cmd_opts):
 
         'uiux_separator_appearance': OptionInfo("<h2>Appearance</h2>", "", gr.HTML),
         "uiux_grid_image_size": OptionInfo(150, "Grid image size", gr.Slider, {"minimum": 64, "maximum": 1024, "step": 1}),
-        "uiux_panel_min_width": OptionInfo(35, "Panel minimum width", gr.Number),
         "uiux_hide_legacy": OptionInfo(True, "Hide legacy tabs"),
         "uiux_persist_layout": OptionInfo(True, "Persist UI layout"),
         "uiux_no_slider_layout": OptionInfo(False, "Hide input range sliders"),